+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Eliminated the need for duplicate `init` call for non-react applications that use pre-built UI. [See the issue](https://github.com/supertokens/supertokens-auth-react/issues/616)
 -   Split/separate pre-built UI components from its recipe to reduce bundle sizes for apps that do not use pre-built UI
+-   Added thirdparty login with popup window example
 
 ### Migration